changed favoriteResponse variables to expose to client, added findFavoritesByUser_id to repository, updated methods in FavoriteService

+++ src/main/java/se/java/security/services/FavoriteService.java
@@ -1,5 +1,7 @@
 package se.java.security.services;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import se.java.security.dto.FavoriteDTO;
 import se.java.security.dto.FavoriteResponse;
@@ -17,8 +19,11 @@ import java.util.stream.Collectors;
 @Service
 public class FavoriteService {
 
+    @Autowired
     private final FavoriteRepository favoriteRepository;
+    @Autowired
     private final UserRepository userRepository;
+    @Autowired
     private final ListingRepository listingRepository;
 
     public FavoriteService(FavoriteRepository favoriteRepository, UserRepository userRepository, ListingRepository listingRepository) {
@@ -30,10 +35,21 @@ public class FavoriteService {
     private FavoriteResponse convertToDTO(Favorite favorite) {
         FavoriteResponse favoriteResponse = new FavoriteResponse();
 
-        favoriteResponse.setUserId(favorite.getUserId().getId());
+        favoriteResponse.setUsername(favorite.getUserId().getId());
+        favoriteResponse.setEmail(favorite.getUserId().getEmail());
 
         favoriteResponse.getFavoritedListingsIds().add(favorite.getListingId().getId());
+
         return favoriteResponse;
+    }
+
+    private static List<FavoriteResponse> getFavoriteResponses(List<Favorite> favorites) {
+        List<FavoriteResponse> favoriteResponses = favorites.stream()
+                .map(fav -> new FavoriteResponse(fav.getUserId().getUsername(),
+                        fav.getUserId().getEmail(),
+                        Collections.singletonList(fav.getListingId().getId())))
+                .toList();
+        return favoriteResponses;
     }
 
     // create a new favorite object
@@ -56,25 +72,31 @@ public class FavoriteService {
         return favoriteRepository.save(newFavorite);
     }
 
-    public List<Favorite> getAllFavorites() {
+    //public favorite response
+    public List<FavoriteResponse> getAllFavorites() {
+        // check that there are favorite objects present
         if (favoriteRepository.findAll().isEmpty()) {
             throw new ResourceNotFoundException("No favorites found");
         }
+
         List<Favorite> favorites = favoriteRepository.findAll();
 
-        return favorites;
+        List<FavoriteResponse> favoriteResponses = getFavoriteResponses(favorites);
+
+        return favoriteResponses;
     }
 
     public List<FavoriteResponse> getUserFavorites(String userId) {
+        // check that the user id exists
         if(!userRepository.existsById(userId)) {
             throw new ResourceNotFoundException("User not found");
         }
 
-        Optional<Favorite> favorites = favoriteRepository.findByUserId(userId);
+        List<Favorite> favorites = favoriteRepository.findFavoritesByUserId_Id(userId);
 
-        return favorites.stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
+        List<FavoriteResponse> favoriteResponses = getFavoriteResponses(favorites);
+
+        return favoriteResponses;
     }
 
     public Optional<Favorite> getSpecificFavorite(String id) {
@@ -85,8 +107,9 @@ public class FavoriteService {
     }
 
     public void deleteFavorite(String id) {
-        Favorite favorite = favoriteRepository.findById(id)
-                        .orElseThrow(() -> new ResourceNotFoundException("Favorite not found with id: " + id));
+        if (!favoriteRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Favorite not found with id: " + id);
+        }
         favoriteRepository.deleteById(id);
     }
 }
